[UPDATE] home라우트 product 라우트에 통합 진행
