+++ routes/productRouter.js
@@ -3,7 +3,8 @@ const productController =require('../controllers/productController');
 
 const router = express.Router();
 
-router.get('/home', productController.getProducts);
+router.get('/all', productController.getProducts);
+router.get('/main', productController.getMain);
 
 module.exports={
     router
